Fixed sticky side header, styling changes
