+++ src/app/page.tsx
@@ -1,10 +1,10 @@
-import Contact from "@/components/Contact";
+import Header from "@/components/Header";
 import Portfolio from "@/components/Portfolio";
 
 export default function Home() {
   return (
-    <div className="flex items-center justify-center h-screen dark:bg-gray-900">
-      <Contact />
+    <div className="lg:flex max-w-screen-xl mx-auto lg:gap-4 leading-relaxed selection:bg-indigo-400">
+      <Header />
       <Portfolio />
     </div>
   );
